+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version 1.31 (not yet released)
 
+* Fixed backwards compatibility with joins on the edit screen.
+
 # Version 1.30
 
 * Fixed issue with relative date and required filters.